+++ src/Libplanet.Net/Messages/BlockHashesMsg.cs
@@ -8,18 +8,9 @@ namespace Libplanet.Net.Messages
 {
     internal class BlockHashesMsg : MessageContent
     {
-        public BlockHashesMsg(long? startIndex, IEnumerable<BlockHash> hashes)
+        public BlockHashesMsg(IEnumerable<BlockHash> hashes)
         {
-            StartIndex = startIndex;
             Hashes = hashes.ToList();
-
-            if (StartIndex is null == Hashes.Any())
-            {
-                throw new ArgumentException(
-                    "The startIndex can be null iff hashes are empty.",
-                    nameof(startIndex)
-                );
-            }
         }
 
         public BlockHashesMsg(byte[][] dataFrames)
@@ -28,8 +19,7 @@ namespace Libplanet.Net.Messages
             var hashes = new List<BlockHash>(hashCount);
             if (hashCount > 0)
             {
-                StartIndex = BitConverter.ToInt64(dataFrames[1], 0);
-                for (int i = 2, end = hashCount + 2; i < end; i++)
+                for (int i = 1, end = hashCount + 1; i < end; i++)
                 {
                     hashes.Add(new BlockHash(dataFrames[i]));
                 }
@@ -37,12 +27,6 @@ namespace Libplanet.Net.Messages
 
             Hashes = hashes;
         }
-
-        /// <summary>
-        /// The block index of the first hash in <see cref="Hashes"/>.
-        /// It is <see langword="null"/> iff <see cref="Hashes"/> are empty.
-        /// </summary>
-        public long? StartIndex { get; }
 
         /// <summary>
         /// The continuous block hashes, from the lowest index to the highest index.
@@ -58,12 +42,7 @@ namespace Libplanet.Net.Messages
             {
                 var frames = new List<byte[]>();
                 frames.Add(BitConverter.GetBytes(Hashes.Count()));
-                if (StartIndex is { } offset)
-                {
-                    frames.Add(BitConverter.GetBytes(offset));
-                    frames.AddRange(Hashes.Select(hash => hash.ToByteArray()));
-                }
-
+                frames.AddRange(Hashes.Select(hash => hash.ToByteArray()));
                 return frames;
             }
         }