+++ src/Libplanet/Blockchain/BlockChain.cs
@@ -662,11 +662,10 @@ namespace Libplanet.Blockchain
         /// <param name="locator">The <see cref="BlockLocator"/> to find the branching point
         /// from.</param>
         /// <param name="count">The Maximum number of <see cref="BlockHash"/>es to return.</param>
-        /// <returns>A tuple of the index of the branch point and <see cref="BlockHash"/>es
-        /// including the branch point <see cref="BlockHash"/>.  If no branch point is found,
-        /// returns a tuple of <see langword="null"/> and an empty array of
-        /// <see cref="BlockHash"/>es.</returns>
-        public Tuple<long?, IReadOnlyList<BlockHash>> FindNextHashes(
+        /// <returns>A list of <see cref="BlockHash"/>es including
+        /// the branch point <see cref="BlockHash"/>.  If no branch point is found,
+        /// returns an empty list of <see cref="BlockHash"/>es.</returns>
+        public IReadOnlyList<BlockHash> FindNextHashes(
             BlockLocator locator,
             int count = 500)
         {
@@ -675,12 +674,12 @@ namespace Libplanet.Blockchain
 
             if (!(FindBranchpoint(locator) is { } branchpoint))
             {
-                return new Tuple<long?, IReadOnlyList<BlockHash>>(null, Array.Empty<BlockHash>());
+                return Array.Empty<BlockHash>();
             }
 
             if (!(Store.GetBlockIndex(branchpoint) is { } branchpointIndex))
             {
-                return new Tuple<long?, IReadOnlyList<BlockHash>>(null, Array.Empty<BlockHash>());
+                return Array.Empty<BlockHash>();
             }
 
             var result = new List<BlockHash>();
@@ -705,7 +704,7 @@ namespace Libplanet.Blockchain
                     Store.ListChainIds().Count(),
                     stopwatch.ElapsedMilliseconds);
 
-            return new Tuple<long?, IReadOnlyList<BlockHash>>(branchpointIndex, result);
+            return result;
         }
 
         /// <summary>
@@ -1190,34 +1189,19 @@ namespace Libplanet.Blockchain
         /// <see langword="null"/>.</returns>
         internal BlockHash? FindBranchpoint(BlockLocator locator)
         {
-            try
+            if (ContainsBlock(locator.Hash))
             {
-                _rwlock.EnterReadLock();
-
                 _logger.Debug(
-                    "Finding a branchpoint with locator [{LocatorHead}]",
+                    "Found a branchpoint with locator [{LocatorHead}]: {Hash}",
+                    locator.Hash,
                     locator.Hash);
-                BlockHash hash = locator.Hash;
-                if (_blocks.ContainsKey(hash)
-                    && _blocks[hash] is Block block
-                    && hash.Equals(Store.IndexBlockHash(Id, block.Index)))
-                {
-                    _logger.Debug(
-                        "Found a branchpoint with locator [{LocatorHead}]: {Hash}",
-                        locator.Hash,
-                        hash);
-                    return hash;
-                }
+                return locator.Hash;
+            }
 
-                _logger.Debug(
-                    "Failed to find a branchpoint locator [{LocatorHead}]",
-                    locator.Hash);
-                return null;
-            }
-            finally
-            {
-                _rwlock.ExitReadLock();
-            }
+            _logger.Debug(
+                "Failed to find a branchpoint locator [{LocatorHead}]",
+                locator.Hash);
+            return null;
         }
 
         /// <summary>