+++ test/Libplanet.Net.Tests/Messages/BlockHashesTest.cs
@@ -20,22 +20,10 @@ namespace Libplanet.Net.Tests.Messages
         }
 
         [Fact]
-        public void Constructor()
-        {
-            Assert.Throws<ArgumentException>(() =>
-                new BlockHashesMsg(null, new[] { default(BlockHash) })
-            );
-            Assert.Throws<ArgumentException>(() =>
-                new BlockHashesMsg(123, new BlockHash[0])
-            );
-        }
-
-        [Fact]
         public void Decode()
         {
             BlockHash[] blockHashes = GenerateRandomBlockHashes(100L).ToArray();
-            var msg = new BlockHashesMsg(123, blockHashes);
-            Assert.Equal(123, msg.StartIndex);
+            var msg = new BlockHashesMsg(blockHashes);
             Assert.Equal(blockHashes, msg.Hashes);
             var privateKey = new PrivateKey();
             AppProtocolVersion apv = AppProtocolVersion.Sign(privateKey, 3);
@@ -48,7 +36,6 @@ namespace Libplanet.Net.Tests.Messages
                 peer,
                 DateTimeOffset.UtcNow);
             BlockHashesMsg restored = (BlockHashesMsg)messageCodec.Decode(encoded, true).Content;
-            Assert.Equal(msg.StartIndex, restored.StartIndex);
             Assert.Equal(msg.Hashes, restored.Hashes);
         }