+++ CHANGES.md
@@ -25,8 +25,13 @@ To be released.
  -  (Libplanet.Net) Changed `BlockHashDownloadState` and `BlockDownloadState`
     to be `Obsolete`.  [[#3943]]
  -  Removed `Fork()` method from `BlockChain`.  [[#3948]]
+ -  Changed the return type for `BlockChain.FindNextHashes()`
+    to `IReadOnlyList<BlockHash>`.  [[#3949]]
 
 ### Backward-incompatible network protocol changes
+
+ -  (Libplanet.Net) Changed the encoding for `GetBlockHashesMsg` and
+    `BlockHashesMsg`.  [[#3949]]
 
 ### Backward-incompatible storage format changes
 
@@ -57,6 +62,7 @@ To be released.
 [#3942]: https://github.com/planetarium/libplanet/pull/3942
 [#3943]: https://github.com/planetarium/libplanet/pull/3943
 [#3948]: https://github.com/planetarium/libplanet/pull/3948
+[#3949]: https://github.com/planetarium/libplanet/pull/3949
 
 
 Version 5.2.2